moving auth module to modules folder
