+++ lib/app/modules/auth/domain/entities/login_credentials.dart
