+++ lib/app/modules/auth/domain/entities/user_info.dart
