+++ lib/app/modules/auth/domain/errors/errors.dart
