+++ lib/app/modules/auth/external/datasources/firebase/errors/errors.dart
