+++ lib/app/modules/auth/infra/models/user_model.dart
@@ -1,0 +1,15 @@
+import 'package:market_lists/app/modules/auth/domain/entities/user.dart';
+
+class UserModel extends User {
+  UserModel({
+    super.id,
+    required super.name,
+    super.email,
+    super.phone,
+    super.imageUrl,
+    super.createdAt,
+    super.updatedAt,
+  });
+
+  User toUserImpl() => this;
+}
