+++ lib/app/modules/auth/presenter/bloc/auth_event.dart
