+++ lib/app/modules/auth/presenter/bloc/auth_state.dart
